% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMPSH-package.R
\docType{package}
\name{RMPSH-package}
\alias{RMPSH}
\alias{RMPSH-package}
\title{RMPSH: Recursive Modified Pattern Search on Hyper-Rectangle}
\description{
Optimization of any Black-Box/Non-Convex Function on Hyper-Rectangular Parameter Space. It uses a Variation of Pattern Search Technique. Described in the paper : Das (2016) <arXiv:1604.08616> .
}
\author{
\strong{Maintainer}: Priyam Das \email{pdas@ncsu.edu}

Authors:
\itemize{
  \item Debsurya De \email{debsurya001@gmail.com}
}

}
\keyword{internal}
