% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMPSH-package.R
\docType{package}
\name{RMPSH-package}
\alias{RMPSH}
\alias{RMPSH-package}
\title{RMPSH: Recursive Modified Pattern Search on Hyper-rectangle}
\description{
RMPSH is developped for optimizing any blackbox/non-convex function on hyper-rectangular parameter space.
}
\author{
\strong{Maintainer}: Priyam Das \email{pdas@ncsu.edu}

}
\keyword{internal}
